i2c: tegra: add option to have interruptable i2c transfer
There is need to not have delay on tx fifo to be fill after writing the
packet header to avoid clock stretching from maser due to interruptable
filling of TxFifo. Due to this requirements, the locking is used to avoid
interrupt. But there is some other device which can allow this and so do not
need to disable interrupt and hence lock can be remove.

Add option to select this from the configuration to match with bus expectation.

bug 200036230

Change-Id: Ia13124acec0d4acd36779aefc594946c368dc59f
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/562925